web/requirements.txt
django
djangorestframework
dj_database_url
python-decouple
psycopg2
unipath
gunicorn
